FIKS: Kall for å lage beregningsgrunnlag GUI dto går via BeregningTjeneste
