+++ behandlingsprosess/src/main/java/no/nav/foreldrepenger/behandling/steg/beregningsgrunnlag/VurderRefusjonBeregningsgrunnlagSteg.java
@@ -15,6 +15,8 @@ import no.nav.foreldrepenger.behandlingslager.behandling.BehandlingStegType;
 import no.nav.foreldrepenger.behandlingslager.behandling.repository.BehandlingRepository;
 import no.nav.foreldrepenger.behandlingslager.fagsak.FagsakYtelseType;
 import no.nav.foreldrepenger.domene.mappers.BeregningAksjonspunktResultatMapper;
+import no.nav.foreldrepenger.domene.mappers.til_kalkulator.BeregningsgrunnlagInputFelles;
+import no.nav.foreldrepenger.domene.mappers.til_kalkulator.BeregningsgrunnlagInputProvider;
 import no.nav.foreldrepenger.domene.prosess.BeregningsgrunnlagKopierOgLagreTjeneste;
 
 @BehandlingStegRef(BehandlingStegType.VURDER_REF_BERGRUNN)
