+++ behandlingsprosess/src/test/java/no/nav/foreldrepenger/behandling/steg/beregningsgrunnlag/ForeslåBeregningsgrunnlagStegTest.java
@@ -9,6 +9,7 @@ import java.time.LocalDate;
 import java.util.List;
 import java.util.Set;
 
+import no.nav.foreldrepenger.domene.mappers.til_kalkulator.BeregningsgrunnlagInputProvider;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
