+++ domenetjenester/beregningsgrunnlag/src/main/java/no/nav/foreldrepenger/domene/mappers/til_kalkulator/BeregningsgrunnlagInputFelles.java
@@ -1,4 +1,4 @@
-package no.nav.foreldrepenger.behandling.steg.beregningsgrunnlag;
+package no.nav.foreldrepenger.domene.mappers.til_kalkulator;
 
 import java.util.List;
 import java.util.Objects;
